refactor: decouple package filters
... and improve mocks for package filters.
