+++ src/filters/__mocks__/__common__.ts
@@ -1,0 +1,5 @@
+const filter = () => false;
+
+export {
+  filter,
+};
