+++ src/filters/__mocks__/glob.ts
@@ -1,0 +1,7 @@
+import { filter } from "./__common__";
+
+const NewGlobFilter = jest.fn()
+  .mockReturnValue(filter)
+  .mockName("filters.NewGlobFilter");
+
+export default NewGlobFilter;
