+++ src/filters/__mocks__/pr-files.ts
@@ -1,5 +1,7 @@
+import { filter } from "./__common__";
+
 const NewPrFilesFilter = jest.fn()
-  .mockReturnValue([() => false, null])
+  .mockReturnValue([filter, null])
   .mockName("filters.NewPrFilesFilter");
 
 export default NewPrFilesFilter;
