+++ src/filters/__mocks__/pushed-files.ts
@@ -1,5 +1,7 @@
+import { filter } from "./__common__";
+
 const NewPushedFilesFilter = jest.fn()
-  .mockReturnValue([() => false, null])
+  .mockReturnValue([filter, null])
   .mockName("filters.NewPushedFilesFilter");
 
 export default NewPushedFilesFilter;
