+++ src/filters/types.d.ts
@@ -1,5 +1,8 @@
+import type { error } from "../types";
+
 type FilterFn = (filepath: string) => boolean;
 
 export type {
+  error,
   FilterFn,
 };
